structure e2e tests
